Extract initial executor count to utils class
